12/28/20

+++ dataset/voc_detector_collate.py
@@ -5,7 +5,9 @@ import numpy as np
 def voc_detector_co(batch):
     targets = []
     imgs = []
+    #from IPython import embed;embed()
     for sample in batch:
         imgs.append(sample[0])
-        targets.append((torch.FloatTensor(sample[1][0]),torch.FloatTensor(sample[1][1])))
+        #from IPython import embed;embed()
+        targets.append((sample[1][0].to(torch.float32),sample[1][1]))
     return torch.stack(imgs, 0), targets
